BestInferiorImprovement rule
In its current implementation, this rule does not produce meaningful
results. It just gets stuck in a cycle between the best and best-1
result. The reason is that this rule gets the best result from the given
Neighbourhood, and excludes the current result from being picked again.
Just excluding the current result is not sufficient, a tabu queue is
needed.